update rpcserver implementation of the new rpc services
